requirements.txt
fastapi==0.95.2
starlette==0.27.0
uvicorn[standard]==0.29.0
pandas
mysql-connector-python==8.3.0
python-dotenv==1.0.1
python-multipart==0.0.6
pyspark==3.4.1
pytest
httpx==0.23.3